requirements.txt
pur
django
openai
python-dotenv
python-telegram-bot
pytz
requests
uvicorn
tiktoken
django-import-export